Generate UI input widget

Generate UI input widget

generateUIwidgets(
  colname,
  mode = c("add", "edit"),
  tab = c("Patient", "Sample", "Mutation"),
  data = NULL,
  selected_row = NULL,
  patientIDs = NULL,
  sampleIDs = NULL
)

Arguments

colname
A character string - the name of the column, that will be the label of the input

mode
"add" or "edit" - whether to use existing values or not

tab
"Patient", "Sample" - The used tab; sets the html id prefix of the input

data
A data.frame.

selected_row
A number indicating the row number of the selected row in the data.frame.

patientIDs
Vector of patient IDs used for drop down menu of the PATIENT_ID column

Value

A sanitized string.

Examples

cbpManager:::generateUIwidgets(colname = "attribute", mode = "add", tab = "Patient")
#> <div class="row">
#>   <div class="col-sm-8">
#>     <div class="form-group shiny-input-container">
#>       <label class="control-label" id="addPatientInput_attribute-label" for="addPatientInput_attribute">attribute</label>
#>       <input id="addPatientInput_attribute" type="text" class="form-control" value=""/>
#>     </div>
#>   </div>
#> </div>